Refactored tree sampling method and added RELEASE build in CMake.

+++ src/Trees/TreeParser.cpp
@@ -22,7 +22,7 @@ inline std::string& IO::cleanTreeString(std::string &tree_string) {
 std::pair<std::string, std::string> IO::splitBranchString(std::string branch_string) {
   std::pair<std::string, std::string> b;
   int tree_depth = 0;
-  for (int position = 0; position < branch_string.size(); position++) {
+  for (unsigned int position = 0; position < branch_string.size(); position++) {
     char character = branch_string.at(position);
     //std::cout << "Char: " << character << " Depth: " << tree_depth << std::endl;
     if (character == '(')
@@ -50,7 +50,7 @@ node_data IO::deconstructNodeString(std::string node_string) {
 
   std::string name;
   std::pair<std::string, std::string> branch_strings;
-  if (last_parens_position == string::npos) {
+  if (last_parens_position == (signed int)string::npos) {
     // Tip node.
     name = node_string.substr(0, last_colon_position);
     branch_strings = std::make_pair("", "");
